LPD-94145 Extract the AI Assistant chat state into a useAIChat hook
